refactor(tests): unify cold-start readiness gates into one awaitEdgeReady primitive (ADR 0127, #1721)
Three integration-harness probes hit the same Cloudflare cold-start signature across
two lanes (/fate/live #1689, /api/auth/sign-up/email #1717, /api/health on #1720), each
closed with its own point-fix — drifting into parallel copies of one idea. Generalize
into ONE shared primitive and route every probe through it.

- New `_edge-ready.ts`: the ONE typed placeholder-404 signal
  (`CloudflarePlaceholder404Error` + `isCloudflarePlaceholder404` + `edgeFetch`) and the
  ONE bounded readiness poll `awaitEdgeReady(send, ready, {deadlineMs, pollMs})`
  (ready may be async). This IS the #1689/#1717 mechanism, made the single home.
- `_harness.ts`: `openSse`/`liveControl`/`signUp` route through `awaitEdgeReady`; the
  detectors move to the shared module (`req` imports them).
- `_integration.ts`: `warmLiveDO`/`awaitWorkerReady`/`warmFateRead` re-expressed over
  `awaitEdgeReady` + `edgeFetch`; the per-probe tagged sentinels + `Schedule.spaced`
  loops retired. `awaitWorkerReady` no longer needs an `HttpClient`.
- Tests: the two point-fix files (_fate-live-readiness #1690, _auth-signup-readiness
  #1720) fold into `_edge-ready.unit.test.ts`, re-expressed over the primitive and
  covering all three probes + the load-bearing invariants (placeholder rides the budget;
  abort / genuine 4xx / 422-fallback / terminal worker JSON 4xx fast-fail).

Invariants preserved: CF placeholder-404 / cold-worker rides the bounded budget for
every probe; only that signal rides it (every other throw fast-fails); /fate/live SSE,
signup user-minting, and /api/health worker-ready behave identically.

Fixes #1721
